Refactored service setup to Docker
Removed the existing serve.sh script and replaced it with a Docker based service setup. Created Dockerfile, docker-compose file to enable running the service in a Docker container. New scripts to serve the Jekyll site locally both with and without Docker have been added. Ensured the versions of the gems match the ones in our Gemfile